# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Berita Documentation",
        "description": "Dokumentasi API untuk sistem manajemen berita",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/email/verify/{id}/{hash}": {
            "get": {
                "tags": [
                    "Verification"
                ],
                "summary": "Verifikasi Email User (Klik Link dari Email)",
                "description": "Endpoint ini diakses saat user mengklik link di email. URL ini memiliki signature keamanan.",
                "operationId": "6ed540fa14ef7f273aa29e8927d30c69",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID User",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "hash",
                        "in": "path",
                        "description": "Hash unik email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Email berhasil diverifikasi"
                    },
                    "400": {
                        "description": "Link tidak valid atau kadaluarsa"
                    }
                }
            }
        },
        "/api/email/resend": {
            "post": {
                "tags": [
                    "Verification"
                ],
                "summary": "Kirim Ulang Link Verifikasi",
                "operationId": "43852a3070150f6782a13d075b970da4",
                "responses": {
                    "200": {
                        "description": "Link dikirim ulang"
                    },
                    "400": {
                        "description": "Email sudah diverifikasi"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter your Bearer token",
                "bearerFormat": "JWT"
            }
        }
    },
...
}